+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [v1.0.2](https://github.com/higlass/higlass-python/compare/v1.0.2...v1.0.1)
 
 - Calculate file pointer hash for track uids for tileset tracks
+- Fix view copy behavior to preserve specific plugin track class vars
 
 ## [v1.0.1](https://github.com/higlass/higlass-python/compare/v1.0.1...v1.0.0)